# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: POS Platform API
  version: 0.1.0
  description: >
    REST API for the POS platform providing checkout, backoffice, and
    synchronization capabilities. The API is designed to be idempotent, secure,
    and ready for hybrid deployments.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/products:
    get:
      tags:
        - Catalog
      summary: List products
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Products ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Catalog
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWrite'
      responses:
        '201':
          description: Product created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/products/{id}:
    get:
      tags:
        - Catalog
      summary: Get product
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    patch:
      tags:
        - Catalog
      summary: Update product
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductPatch'
      responses:
        '200':
          description: Product updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '204':
          description: No fields changed.
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/v1/roles:
    get:
      tags:
        - Administration
      summary: List roles
      responses:
        '200':
          description: Available roles ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Administration
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreate'
      responses:
        '201':
          description: Created role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/v1/roles/{id}:
    get:
      tags:
        - Administration
      summary: Get role
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Role details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SearchQuery:
      name: search
      in: query
      description: Case insensitive partial match on name, legacy code, or identifier.
      schema:
        type: string
    LimitQuery:
      name: limit
      in: query
      description: Maximum number of records to return (default 100, max 500).
      schema:
        type: integer
        minimum: 1
        maximum: 500
  responses:
    ValidationError:
      description: Request payload failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: resource not found
    ConflictError:
      description: Unique constraint violation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: conflict
    UnauthorizedError:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: unauthorized
    ForbiddenError:
      description: Authenticated user lacks permission for the requested action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: forbidden
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
        barcode:
          type: string
          nullable: true
        group_id:
          type: string
          format: uuid
          nullable: true
        reference:
          type: string
          nullable: true
        min_stock:
          type: number
          format: float
          nullable: true
        price_cash:
          type: number
          format: float
          nullable: true
        price_base:
          type: number
          format: float
          nullable: true
      required:
        - id
        - name
    ProductWrite:
      type: object
      properties:
        legacy_code:
          type: string
          nullable: true
        name:
          type: string
        barcode:
          type: string
          nullable: true
        group_id:
          type: string
          format: uuid
          nullable: true
        reference:
          type: string
          nullable: true
        min_stock:
          type: number
          format: float
          nullable: true
        price_cash:
          type: number
          format: float
          nullable: true
        price_base:
          type: number
          format: float
          nullable: true
      required:
        - name
    ProductPatch:
      allOf:
        - $ref: '#/components/schemas/ProductWrite'
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        permissions:
          type: array
          items:
            type: string
        discountLimit:
          type: number
          format: float
      required:
        - id
        - name
        - permissions
        - discountLimit
    RoleCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        permissions:
          type: array
          items:
            type: string
        discountLimit:
          type: number
          format: float
          nullable: true
      required:
        - name
        - permissions
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          nullable: true
      required:
        - message
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: validation_failed
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items:
                  type: string
              message:
                type: string
              code:
                type: string
            required:
              - path
              - message
              - code
      required:
        - message
        - issues
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT